refactor: loading the configuration is required by multiple controllers
